+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 
-### NEXT
+### 3.13.14
 
 * Avoid modification of user input data ([PR #1285](https://github.com/versatica/mediasoup/pull/1285)).
 * `ListenInfo`: Add transport socket flags ([PR #1291](https://github.com/versatica/mediasoup/pull/1291)).